Accept a timeout on the collection rerank macro
